backend/requirements.txt
Django==4.0
django-filter==23.2
djangorestframework==3.14.0
django-cors-headers
yt-dlp